Conflicts:
	AndroidManifest.xml
	src/org/geometerplus/android/fbreader/library/FileManager.java

+++ src/org/geometerplus/fbreader/library/BooksDatabase.java
@@ -49,7 +49,7 @@ public abstract class BooksDatabase {
 	protected void addTag(Book book, Tag tag) {
 		book.addTagWithNoCheck(tag);
 	}
-	protected void setSeriesInfo(Book book, String series, long index) {
+	protected void setSeriesInfo(Book book, String series, float index) {
 		book.setSeriesInfoWithNoCheck(series, index);
 	}
